Add NYX_OS_INFO_WEBOS_RELEASE_PLATFORMCODE
:Release Notes:
Add NYX_OS_INFO_WEBOS_RELEASE_PLATFORMCODE represents the version of
webOS Platform codename. "webos_release" is the combination of:

 ${WEBOS_DISTRO_PLATFORM_CODENAME}-${WEBOS_DISTRO_VERSION_CODENAME}

So NYX_OS_INFO_WEBOS_RELEASE_PLATFORMCODE is the value that can
programmatically evaluate it and check an upgrade or downgrade
on Platform perspective.

:Detailed Notes:
NYX_OS_INFO_WEBOS_RELEASE_PLATFORMCODE has the type of value
which can be possible to be evaluated programmatically and higher
value indicates more recent version.

:Testing Performed:
Tested if OSInfo query and its results

:QA Notes:
N/A

:Issues Addressed:
[TVPLAT-75459] Review for OSInfo query results
[TVPLAT-74992] Add supports platform version API for Nyx OSInfo

Change-Id: Ia06c4d45888e0fc47dde323eee7dcd4d659129ad
Reviewed-on: http://gpro.lge.com/c/webos/nyx-utils/+/311576
Reviewed-by: Commit Msg Checker <commit_msg@lge.com>
Reviewed-by: Ban Word Checker <ban_word@lge.com>
Reviewed-by: <madrol10.stephen@lge.com>
Reviewed-by: Kumaran S <kumaran.s@lgepartner.com>
Reviewed-by: Yogish S <yogish.s@lge.com>
Tested-by: Kumaran S <kumaran.s@lgepartner.com>
Tested-by: Yogish S <yogish.s@lge.com>