Add NYX_OS_INFO_WEBOS_RELEASE_PLATFORMCODE
:Release Notes:
Add NYX_OS_INFO_WEBOS_RELEASE_PLATFORMCODE represents the version of
webOS Platform codename. webos_release is the combination of:

 -

So NYX_OS_INFO_WEBOS_RELEASE_PLATFORMCODE is the value that can
programmatically evaluate it and check an upgrade or downgrade
on Platform perspective.

:Detailed Notes:
NYX_OS_INFO_WEBOS_RELEASE_PLATFORMCODE has the type of value
which can be possible to be evaluated programmatically and higher
value indicates more recent version.

:Testing Performed:
Tested if OSInfo query and its results

:QA Notes:
N/A

:Issues Addressed:
[TVPLAT-75459] Review for OSInfo query results
[TVPLAT-74992] Add supports platform version API for Nyx OSInfo

Change-Id: I985e9c255c7c66e793f754f4a246118968bc01cb
Reviewed-on: http://gpro.lge.com/c/webos/nyx-utils/+/312668
Reviewed-by: Commit Msg Checker <commit_msg@lge.com>
Reviewed-by: Ban Word Checker <ban_word@lge.com>
Reviewed-by: Kumaran S <kumaran.s@lgepartner.com>
Reviewed-by: Yogish S <yogish.s@lge.com>
Tested-by: Kumaran S <kumaran.s@lgepartner.com>
Tested-by: Yogish S <yogish.s@lge.com>